Revert "feat: trpc security"
This reverts commit 275a4a397062bada4b9f32349eea1ec55df78ed3.
