+++ backend/core-api/src/main.ts
@@ -1,6 +1,6 @@
 import cookieParser from 'cookie-parser';
 import cors from 'cors';
-import { isDev, setupTRPCRoute } from 'erxes-api-shared/utils';
+import { isDev } from 'erxes-api-shared/utils';
 import express from 'express';
 import * as http from 'http';
 import { appRouter } from '~/init-trpc';
@@ -78,16 +78,19 @@ app.get('/subscriptionPlugin.js', fileLimiter, async (_req, res) => {
   res.sendFile(apolloSubscriptionPath);
 });
 
-setupTRPCRoute(app, {
-  router: appRouter,
-  createContext: async (subdomain, context) => {
-    const models = await generateModels(subdomain);
+app.use(
+  '/trpc',
+  trpcExpress.createExpressMiddleware({
+    router: appRouter,
+    createContext: createTRPCContext(async (subdomain, context) => {
+      const models = await generateModels(subdomain);
 
-    context.models = models;
+      context.models = models;
 
-    return context;
-  },
-});
+      return context;
+    }),
+  }),
+);
 
 app.get('/health', async (_req, res) => {
   res.end('ok');
