+++ backend/erxes-api-shared/src/utils/saas/saas-mongo-connection.ts
@@ -149,7 +149,6 @@ export const getSaasOrganizationDetail = async ({
   subdomain,
 }: {
   subdomain: string;
-  models?: any;
 }) => {
   await getSaasCoreConnection();
