+++ backend/plugins/frontline_api/src/init-trpc.ts
@@ -23,7 +23,7 @@ export const appRouter = t.mergeRouters(
         .input(
           z.object({
             moduleType: z.string(),
-            collectionType: z.string(),
+            collectionType: z.string().optional(),
             segmentId: z.string().optional(),
             usageType: z.string().optional(),
             config: z.record(z.any()).optional(),
