+++ package.json
@@ -80,7 +80,6 @@
     "graphql-tag": "^2.12.6",
     "graphql-ws": "^5.16.0",
     "handlebars": "^4.7.8",
-    "helmet": "^8.1.0",
     "http-proxy-middleware": "^3.0.3",
     "i18next": "^23.16.4",
     "i18next-browser-languagedetector": "^8.0.0",
